# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TRZWorld with Swagger",
        "description": "L5 Swagger OpenApi",
        "contact": {
            "email": "carlosbelchiorsilva@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/report-infections/report": {
            "post": {
                "tags": [
                    "Report Infection"
                ],
                "summary": "Report a survivor as infected",
                "operationId": "a0d34e6f1f4e623c7694ef40ee5b00f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "infected_survivor_id",
                                    "reporting_survivor_id"
                                ],
                                "properties": {
                                    "infected_survivor_id": {
                                        "description": "ID of the survivor reported as infected",
                                        "type": "integer"
                                    },
                                    "reporting_survivor_id": {
                                        "description": "ID of the survivor reporting the infection",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message indicating successful report of infection",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating internal server error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reports/percentage-infected": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Display the percentage of not infected survivors",
                "operationId": "8ea2f45f3b799007c1ad273562c6a2d0",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "percentage": {
                                            "description": "Percentage of not infected survivors",
                                            "type": "number",
                                            "format": "double"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating internal server error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reports/percentage-not-infected": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Display the percentage of not infected survivors",
                "operationId": "0ae27d9d714c9a659a15b647619ed162",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "percentage": {
                                            "description": "Percentage of not infected survivors",
                                            "type": "number",
                                            "format": "double"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating internal server error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reports/average-items": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Calculate average quantity of items per survivor",
                "operationId": "7421770794268946512d86f149d5b511",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "average_items": {
                                            "description": "Average quantity of items per survivor",
                                            "type": "object",
                                            "additionalProperties": {
                                                "description": "Average quantity of the item per survivor",
                                                "type": "number",
                                                "format": "double"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating internal server error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reports/total-points-lost": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Calculate total points lost due to infected survivors",
                "operationId": "8ac186a775921a55dda8fe644fe5e81a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message indicating total points lost",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No infected survivors found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating no infected survivors found",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating internal server error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}